Add auto-help-updating to ArgParsePlugin
The ArgParsePlugin now can look at the values provided after it for a
particular key and update the help field for the arg with a summary of
all the sources that will be used if the arg is not provided on the
command line.

This can be enabled for all fields by adding update_help=True to the
ArgParsePlugin initializer. Or adding update_help=True to an
individual with_argparse call.

A description field was added to all of the Provider classes to help
populate this information. This means the interface for new providers
has expanded. There is currently not a base class or Protocol defined
for Providers.